LPS-98416 Use 'internal' instead of 'impl' for internal classes
