+++ modules/apps/layout/layout-type-controller/layout-type-controller-test/src/testIntegration/java/com/liferay/layout/type/controller/portlet/internal/test/LayoutTypePortletImplBaseTest.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.layout.type.controller.portlet.impl.test;
+package com.liferay.layout.type.controller.portlet.internal.test;
 
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.model.Group;
